Backend/requirements.txt
fastapi
uvicorn
assemblyai
google-generativeai